minimise mkdir/symlink/other errors causing scratch space deletion

+++ scripts/job.sh
@@ -25,7 +25,7 @@ trap finish EXIT ERR INT TERM
 
 #################################################
 #
-# CHOOSE SAMPLE AND CREATE LINKS TO BAMS
+# CHOOSE SAMPLE AND CREATE LINKS TO LARGE FILES
 #
 #################################################
 
@@ -36,27 +36,28 @@ SAMPLE_ID=${SAMPLE_IDS[SGE_TASK_ID]}
 CTL_BAM=${CTL_FILES[SGE_TASK_ID]}
 TRT_BAM=${TRT_FILES[SGE_TASK_ID]}
 SAMPLE_DIR="$RESULTS_BASE_DIR/$SAMPLE_ID" # compute in individual task; should be a full path to an individual's files with sanitised names
-mkdir $SAMPLE_DIR
 BAM_DIR="$SAMPLE_DIR/bams"
-mkdir $BAM_DIR
 CTL_BAM_LINK=$BAM_DIR/$(basename $CTL_BAM | perl -ne 's/\./_/g; s/_bam$/.bam/; print') # sanitise bam names and link sanitised names to orig files, either in scratch or orig location
 TRT_BAM_LINK=$BAM_DIR/$(basename $TRT_BAM | perl -ne 's/\./_/g; s/_bam$/.bam/; print')
 CTL_BASENAME=$(basename $CTL_BAM_LINK | perl -ne 's/\.bam$//; print') # remove .bam
 TRT_BASENAME=$(basename $TRT_BAM_LINK | perl -ne 's/\.bam$//; print')
 
+[ ! -d $SAMPLE_DIR ] && mkdir $SAMPLE_DIR
+[ ! -d $BAM_DIR    ] && mkdir $BAM_DIR
+
+# remove existing links which may be misconnected
+[ -L $CTL_BAM_LINK     ] && rm -f $CTL_BAM_LINK
+[ -L $CTL_BAM_LINK.bai ] && rm -f $CTL_BAM_LINK.bai
+[ -L $TRT_BAM_LINK     ] && rm -f $TRT_BAM_LINK
+[ -L $TRT_BAM_LINK.bai ] && rm -f $TRT_BAM_LINK.bai
+
 USE_SCRATCH=true
-if [ ! $USE_SCRATCH ]
+if [ $USE_SCRATCH ]
 then
-    # use files directly via symlink - forbidden on UCL HPC
-    ln -s $CTL_BAM $CTL_BAM_LINK
-    ln -s $TRT_BAM $TRT_BAM_LINK
-    ln -s $CTL_BAM.bai $CTL_BAM_LINK.bai
-    ln -s $TRT_BAM.bai $TRT_BAM_LINK.bai
-else
     # use scratch space
     echo "mkdir -p $SCRATCH_DIR"
     mkdir -p $SCRATCH_DIR
-    time rsync -av $CTL_BAM $TRT_BAM $CTL_BAM.bai $TRT_BAM.bai $SCRATCH_DIR
+    time rsync -av $CTL_BAM $TRT_BAM $CTL_BAM.bai $TRT_BAM.bai $SCRATCH_DIR/
     ln -s $SCRATCH_DIR/$(basename $CTL_BAM) $CTL_BAM_LINK
     ln -s $SCRATCH_DIR/$(basename $TRT_BAM) $TRT_BAM_LINK
     ln -s $SCRATCH_DIR/$(basename $CTL_BAM).bai $CTL_BAM_LINK.bai
@@ -65,6 +66,12 @@ else
     TRT_BAM=$SCRATCH_DIR/$(basename $TRT_BAM)
     echo "scratch contents"
     ls -l $SCRATCH_DIR
+else
+    # use files directly via symlink - forbidden on UCL HPC
+    ln -s $CTL_BAM $CTL_BAM_LINK
+    ln -s $TRT_BAM $TRT_BAM_LINK
+    ln -s $CTL_BAM.bai $CTL_BAM_LINK.bai
+    ln -s $TRT_BAM.bai $TRT_BAM_LINK.bai
 fi
 
 
@@ -78,62 +85,49 @@ fi
 ### SET SEQUENZA VARS AND DIRS ###
 
 SEQUENZA_DIR=$SAMPLE_DIR/sequenza
-if [ ! -d $SEQUENZA_DIR ]
-then
-    mkdir $SEQUENZA_DIR
-fi
+[ ! -d $SEQUENZA_DIR ] && mkdir $SEQUENZA_DIR
 
 
 ### CREATE PILEUPS, BINNED SEQZ FILES ###
 
 conda activate sequenza
 
+if [ $USE_SCRATCH ]
+then
+    CTL_PILEUP=$SCRATCH_DIR/ctl.pileup
+    TRT_PILEUP=$SCRATCH_DIR/trt.pileup
+    SEQZ_FILE=$SCRATCH_DIR/binned.seqz.gz
+    [ -f $SEQUENZA_DIR/ctl.pileup     ] && time rsync -av $SEQUENZA_DIR/ctl.pileup     $CTL_PILEUP
+    [ -f $SEQUENZA_DIR/trt.pileup     ] && time rsync -av $SEQUENZA_DIR/trt.pileup     $TRT_PILEUP
+    [ -f $SEQUENZA_DIR/binned.seqz.gz ] && time rsync -av $SEQUENZA_DIR/binned.seqz.gz $SEQZ_FILE
+else
+    CTL_PILEUP=$SEQUENZA_DIR/ctl.pileup
+    TRT_PILEUP=$SEQUENZA_DIR/trt.pileup
+    SEQZ_FILE=$SEQUENZA_DIR/binned.seqz.gz
+fi
+
 if [ ! -f $SEQUENZA_DIR/ctl.pileup ]
 then
-
-    if [ $USE_SCRATCH ]
-    then
-        CTL_PILEUP=$SCRATCH_DIR/ctl.pileup
-    else
-        CTL_PILEUP=$SEQUENZA_DIR/ctl.pileup
-    fi
 
     samtools mpileup --fasta-ref $PPLN_BASE_DIR/genome/$GENOME"_"$CHR_SUFFIX".fa" \
         --positions $PPLN_BASE_DIR/genome/$GENOME"_"$CHR_SUFFIX"_exons.bed" \
         --min-MQ 20 --excl-flags 256 $CTL_BAM \
         -o $CTL_PILEUP # ~68GB bam took 1h12m2s clock time to run, producing file of size 8,818,616,226; <1GB RAM
 
-    if [ $USE_SCRATCH ]
-    then
-        rsync -av $CTL_PILEUP $SEQUENZA_DIR/ctl.pileup
-    fi
+    [ $USE_SCRATCH ] && [ -f $CTL_PILEUP ] && rsync -av $CTL_PILEUP $SEQUENZA_DIR/ctl.pileup
 
-else
-    CTL_PILEUP=$SEQUENZA_DIR/ctl.pileup
 fi
 
 if [ ! -f $SEQUENZA_DIR/trt.pileup ]
 then
-
-    if [ $USE_SCRATCH ]
-    then
-        TRT_PILEUP=$SCRATCH_DIR/trt.pileup
-    else
-        TRT_PILEUP=$SEQUENZA_DIR/trt.pileup
-    fi
 
     samtools mpileup --fasta-ref $PPLN_BASE_DIR/genome/$GENOME"_"$CHR_SUFFIX".fa" \
         --positions $PPLN_BASE_DIR/genome/$GENOME"_"$CHR_SUFFIX"_exons.bed" \
         --min-MQ 20 --excl-flags 256 $TRT_BAM \
         -o $TRT_PILEUP # ~68GB bam took 1h4m18s clock time to run, producing file of size 8,342,805,140; would this be faster if connections were faster?
 
-    if [ $USE_SCRATCH ]
-    then
-        rsync -av $TRT_PILEUP $SEQUENZA_DIR/trt.pileup
-    fi
+    [ $USE_SCRATCH ] && [ -f $TRT_PILEUP ] && rsync -av $TRT_PILEUP $SEQUENZA_DIR/trt.pileup
 
-else
-    TRT_PILEUP=$SEQUENZA_DIR/trt.pileup
 fi
 
 if [ ! -f $SEQUENZA_DIR/binned.seqz.gz ]
@@ -141,7 +135,9 @@ then
     sequenza-utils bam2seqz -p -n $CTL_PILEUP -t $TRT_PILEUP \
         --fasta $PPLN_BASE_DIR/genome/$GENOME"_"$CHR_SUFFIX".fa" \
         -gc $PPLN_BASE_DIR/genome/$GENOME"_"$CHR_SUFFIX".gc50Base.wig.gz" | \
-        sequenza-utils seqz_binning -s - -w 100000 -o $SEQUENZA_DIR/binned.seqz.gz # pileups from two 68GB bams took 29m9s clock time to run, producing file of size 17,025,764
+        sequenza-utils seqz_binning -s - -w 100000 -o $SEQZ_FILE # pileups from two 68GB bams took 29m9s clock time to run, producing file of size 17,025,764
+
+    [ $USE_SCRATCH ] && [ -f $SEQZ_FILE ] && rsync -av $SEQZ_FILE $SEQUENZA_DIR/binned.seqz.gz
 fi
 
 
@@ -182,7 +178,7 @@ then
         ploidy <- confint$max.ploidy;
         cellularity <- confint$max.cellularity;
         cat("\tPloidy\ttumorPurity\ttumorPloidy\n",sample_name,"\t2\t",cellularity,"\t",ploidy,"\n",sep="");
-    ' $SEQUENZA_DIR/binned.seqz.gz $TRT_BASENAME > $SEQUENZA_DIR/purity_ploidy.txt
+    ' $SEQZ_FILE $TRT_BASENAME > $SEQUENZA_DIR/purity_ploidy.txt || true
     #   Ploidy  tumorPurity     tumorPloidy     
     #example_tumor_sorted    2       0.8     1.8     # not sure what Ploidy is for, need tumor* vars for lohhla
 fi
@@ -200,26 +196,21 @@ conda deactivate
 ### SET POLYSOLVER VARS AND DIRS ###
 
 POLYSOLVER_DIR=$SAMPLE_DIR/polysolver
-if [ ! -d $POLYSOLVER_DIR ]
-then
-    mkdir $POLYSOLVER_DIR
-fi
-
-POLYSOLVER_WORK_DIR=$POLYSOLVER_DIR/work
-if [ ! -d "$POLYSOLVER_WORK_DIR" ]
-then
-    mkdir $POLYSOLVER_WORK_DIR
-fi
-
 POLYSOLVER_OUT_DIR=$POLYSOLVER_DIR/results
-if [ ! -d $POLYSOLVER_OUT_DIR ]
-then
-    mkdir $POLYSOLVER_OUT_DIR
-fi
+[ ! -d $POLYSOLVER_DIR ] && mkdir $POLYSOLVER_DIR
+[ ! -d $POLYSOLVER_OUT_DIR ] && mkdir $POLYSOLVER_OUT_DIR
 
-# export vars so accessible to polysolver script
-export HLA_DATA_DIR=$PPLN_BASE_DIR/hla_fasta
-export HLA_FASTA=abc_complete.fasta
+# vars assumed accessible to polysolver script
+HLA_DATA_DIR=$PPLN_BASE_DIR/hla_fasta
+HLA_FASTA=abc_complete.fasta
+
+if [ $USE_SCRATCH ]
+then
+    POLYSOLVER_WORK_DIR=$SCRATCH_DIR/polysolver
+else
+    POLYSOLVER_WORK_DIR=$POLYSOLVER_DIR/work
+fi
+[ ! -d $POLYSOLVER_WORK_DIR ] && mkdir $POLYSOLVER_WORK_DIR
 
 #echo "running pipeline with genome $GENOME; files-list $SAMPLE_METADATA; SGE_TASK_ID $SGE_TASK_ID; pt-id $SAMPLE_ID; control-bam $CTL_BAM; treated-bam $TRT_BAM; bam-dir $BAM_DIR; target-bam $BAM; work-dir $WORK_DIR; out-dir $OUT_DIR"
 
@@ -230,12 +221,12 @@ if [ ! -f $POLYSOLVER_OUT_DIR/winners.hla.txt ]
 then
 
     conda activate polysolver
-    $PPLN_BASE_DIR/polysolver_inst/bin/shell_call_hla_type $CTL_BAM Unknown 1 $GENOME STDFQ 0 $POLYSOLVER_WORK_DIR # slightly modified polysolver script, permitting custom HLA allele library
+    . $PPLN_BASE_DIR/polysolver_inst/bin/shell_call_hla_type $CTL_BAM Unknown 1 $GENOME STDFQ 0 $POLYSOLVER_WORK_DIR || true # slightly modified polysolver script, permitting custom HLA allele library
     conda deactivate
 
     if [ -f $POLYSOLVER_WORK_DIR/winners.hla.txt ]
     then
-        mv $POLYSOLVER_WORK_DIR/winners.hla.txt $POLYSOLVER_OUT_DIR
+        rsync -av $POLYSOLVER_WORK_DIR/winners.hla.txt $POLYSOLVER_OUT_DIR/winners.hla.txt
         rm -r $POLYSOLVER_WORK_DIR/*
     fi
 fi
@@ -256,28 +247,12 @@ TRT_BAM_SIZE=$(stat -c%s "$TRT_BAM")
 TRT_BAM_READ_COUNT=$(($TRT_BAM_SIZE / 80))
 
 LOHHLA_DIR=$SAMPLE_DIR/lohhla
-if [ ! -d $LOHHLA_DIR ]
-then
-    mkdir $LOHHLA_DIR
-fi
-
 LOHHLA_WORK_DIR=$LOHHLA_DIR/work
-if [ ! -d $LOHHLA_DIR/work ]
-then
-    mkdir $LOHHLA_DIR/work
-fi
-
 LOHHLA_OUT_DIR=$LOHHLA_DIR/results
-if [ ! -d $LOHHLA_DIR/results ]
-then
-    mkdir $LOHHLA_DIR/results
-fi
-
-if [ ! -d $LOHHLA_DIR/results/Figures ]
-then
-    mkdir $LOHHLA_DIR/results/Figures
-fi
-#echo "running lohhla with genome $GENOME; pt-id $SAMPLE_ID; out-dir $OUT_DIR"
+[ ! -d $LOHHLA_DIR ] && mkdir $LOHHLA_DIR
+[ ! -d $LOHHLA_DIR/work ] && mkdir $LOHHLA_DIR/work
+[ ! -d $LOHHLA_DIR/results ] && mkdir $LOHHLA_DIR/results
+[ ! -d $LOHHLA_DIR/results/Figures ] && mkdir $LOHHLA_DIR/results/Figures
 
 
 ### RUN LOHHLA ###
@@ -309,7 +284,7 @@ if [ $CTL_BAM_SORTED -eq 0 ]; then
     ln -s $CTL_SORTED_BAM.bai $CTL_BAM_LINK.bai
 fi
 
-# lohhla is fork of slagtermaarten/LOHHLA
+# LOHHLA is fork of bitbucket.org/mcgranahanlab/lohhla
 #echo "$SHELL"
 Rscript $PPLN_BASE_DIR/LOHHLA/LOHHLAscript.R \
   --genomeAssembly=$GENOME \
